refactor: resolve top 5 critical audit issues across architecture
Connection pooling (PooledConnection + ConexaoBD rewrite) eliminates
per-call connection overhead. In-memory cache in AuxiliaresDAO removes
N+1 queries for auxiliary lookups (500+ connections/session → ~5).

double → BigDecimal in 6 models and 5 financial controllers prevents
floating-point rounding in money calculations. StatusPagamento tolerance
now uses BigDecimal("0.01") instead of double literal.

Auth hardening: LoginController uses BCrypt via UsuarioDAO instead of
plaintext SQL comparison. Plaintext fallbacks removed from estorno and
financeiro controllers. DB credentials externalized to db.properties.

AuxiliaresDAO refactored from 35 identical CRUD methods to 5 generic
ones with table whitelist validation. Legacy methods kept as one-line
delegates for compatibility.

UI threading: TelaPrincipal, ListaEncomenda, CadastroViagem, and
InserirEncomenda now load DB data in background Tasks. 38 empty catch
blocks fixed across 13 files with contextual error logging.

+++ src/dao/FreteDAO.java
@@ -88,9 +88,9 @@ public class FreteDAO {
                         frete.setDataEmissao(dataEmissaoSql.toLocalDate());
                     }
                     
-                    frete.setValorNominal(rs.getDouble("valor_nominal"));
-                    frete.setValorDevedor(rs.getDouble("valor_devedor"));
-                    frete.setValorPago(rs.getDouble("valor_pago"));
+                    frete.setValorNominal(rs.getBigDecimal("valor_nominal"));
+                    frete.setValorDevedor(rs.getBigDecimal("valor_devedor"));
+                    frete.setValorPago(rs.getBigDecimal("valor_pago"));
                     frete.setNomeConferente(rs.getString("conferente"));
                     frete.setStatus(rs.getString("status_frete"));
                     frete.setTotalVolumes(rs.getInt("total_volumes"));
